// src/app/domain/product/enums/product_categories.ts
export enum ProductCategory {
    ELECTRONICS = 1,
    CLOTHING = 2,
    HOME = 3,
    BEAUTY = 4,
    SPORTS = 5,
    TOOLS = 6,
    FOOD = 7,
    DRINKS = 8,
    HOUSEHOLD = 9,
    PETS = 10,
    HEALTH = 11,
    OTHER = 99
}

export enum ProductSubCategories {

    // People References
    BABY = 1,
    CHILD = 2,
    WOMEN = 3,
    MEN = 4,
    ADULT = 5,

    // Pets References
    PETS = 6,

    // Accessories References
    ACCESSORIES = 7,

    // Games References
    GARDEN = 8, 
    OUTDOOR = 9,

    // Entertainment References
    VIDEOGAMES = 10,
    MUSIC = 11,
    MULTIMEDIA = 12,
    BOOKS = 13,
    SCHOOL = 14,
    EDUCATION = 15,
    WRITING = 16,

    // Health/Beauty References
    HEALTH = 18,
    FITNESS = 19,
    BEAUTY = 20,
    SKINCARE = 21,
    HAIRCARE = 22,
    NAILCARE = 23,
    MAKEUP = 24,
    FRAGRANCE = 25,
    BODYCARE = 26,
    HANDCARE = 27,
    FOOTCARE = 28,
    EYECARE = 29,
    EARCARE = 30,
    NOSECARE = 31,
    MOUTHCARE = 32,
    TOOTHCARE = 33,
    DENTALCARE = 34,

    // Food/Drinks References
    FOOD = 35,
    DRINKS = 36,
    ALCOHOL = 37,
    BEVERAGES = 38,
    SNACKS = 39,
    CANDY = 40,
    
    // Household References
    HOUSEHOLD = 41,
    CLEANING = 42,
    ORGANIZATION = 43,
    STORAGE = 44,
    DECORATION = 45,
    TOOLS = 46,
    CONSTRUCTION = 47,
    REPAIR = 48,
    OFFICE = 49,

    // Clothing References
    CLOTHING = 50,
    SHOES = 51,
    CLOTHES = 52,
    BAGS = 53,
    JEWELRY = 54,
    WATCHES = 55,
    SUNGLASSES = 56,

    // Other Reference
    OTHER = 99

}